remove "--nbd=nbdkit-no-sa" option parsing
According to the first commit message in this series, remove support for
parsing "--nbd=nbdkit-no-sa".

Systemd socket activation was added to nbdkit in nbdkit commit
7ff39d028c63 ("Add support for socket activation.", 2017-01-31); first
released in nbdkit v1.1.13 (Aug 2017).

Note that there is a non-intuitive change in this patch: in the
test_nbd_servers() function, we remove the grepping for the "LISTEN_PID"
string in the "nbdkit" binary. In other words, we stop verifying whether
"nbdkit" indeed supports socket activation, we just assume it.

The primary reason for this is that this check never actually worked when
the nbdkit binary was exposed from the root directory of an nbdkit build
tree! That binary is actually built from "wrapper.c", and it's only the
"server/nbdkit" binary that matches "LISTEN_PID". However, as "wrapper.c"
explains, "server/nbdkit" would use the system-wide nbdkit plugins, not
the just-built ones, therefore only the wrapper "nbdkit" executable can be
used -- and so the "LISTEN_PID" grep command has always been unable to
deal with that situation. Unless we remove that check too, with the rest
of this patch, virt-p2v refuses to start up -- as no usable nbd server is
found.

This patch permits further simplification, which will be done subsequently
in this series.

Ref: https://listman.redhat.com/archives/libguestfs/2022-March/028475.html
Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>