nbd.c: simplify start_nbdkit()
The previous patch guarantees that start_nbdkit() is only called with a
non-NULL "fds" parameter (and, consequently, that start_nbdkit() does not
use "ipaddr" and "port" for anything beyond logging).

Remove the branch in start_nbdkit() that starts nbdkit without socket
activation, and trim the start_nbdkit() prototype too.

This patch is best viewed with "git show -b" due to the unindentation in
it.

Ref: https://listman.redhat.com/archives/libguestfs/2022-March/028475.html
Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>