rely on Linux for killing nbdkit, when the parent thread exits
We currently track the PIDs of the NBD servers (nbdkit only, at this
point) so that we can kill and reap them in cleanup_data_conns().

cleanup_data_conns() is called from three kinds of contexts:

(1) in start_conversion(), followed immediately by exit (EXIT_FAILURE);

(2) at the end of start_conversion() in GUI mode, where we return to
    start_conversion_thread(), and then the conversion thread -- a
    detached thread -- exits almost immediately;

(3) at the end of start_conversion() in kernel mode. Here we return to
    kernel_conversion(), then main(), and soon exit the process.

As shown above, whenever we intend to kill and reap the nbdkit processes,
in any of these contexts, the parent *thread's* disappearance is also
imminent.

This gives rise to the idea of (a) removing the tracking & explicit
killing of nbdkit processes by PID, and (b) telling the kernel to kill the
nbdkit processes *upon* parent thread disappearance.

And that's what the "--exit-with-parent" option of nbdkit does -- at least
on Linux <https://libguestfs.org/nbdkit-captive.1.html#EXIT-WITH-PARENT>
--, so replace the tracking with "--exit-with-parent".

... I have no clue how to test this patch (or how the patch affects
non-Linux host OSes).

Ref: https://listman.redhat.com/archives/libguestfs/2022-March/028475.html
Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>